requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.1.7
django-appconf==1.0.4
django-googledrive-storage==1.5.2
django-nested-admin==3.3.3
django-nested-inlines==0.1
djangorestframework==3.12.2
drf-yasg==1.20.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
hiredis==2.0.0
httplib2==0.19.0
httplib2shim==0.0.3
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
msgpack==1.0.2
nltk==3.6.1
numpy==1.20.2
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.60.0
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
zope.interface==5.4.0
